Use same text for gavc_download converge_by as remote_file
The whole point of changing the remote_file name was to ensure that the resolved
GAVC download URI would be logged during a chef run. It turns out that the
remote_file name is not logged at all. So, change the provider to use the same
name for converge_by as for the remote_file. Unfortunately, I'm not sure how to
verify the converge_by text using chefspec, so there's no test change.