Add tests for the username and password attributes
In order to make username and password more testable, I eliminated the in-place sub!() on download_uri. Otherwise, that messes with the string that came from the test. The test has to return a download_uri from the Artifactory stub that does not have the username and password in it, but expect a remote_file with a source that does include those values.